// lib/models/types.ts
export type MetodoPago = 'tarjeta' | 'paypal' | 'transferencia';